requirements.txt
imageio==2.33.1
lightning==2.5.1.post0
matplotlib==3.10.6
numpy==2.3.3
pandas==2.3.2
Pillow==11.3.0
PyYAML==6.0.2
PyYAML==6.0
PyYAML==6.0.1
PyYAML==6.0.2
rawpy==0.25.0
skimage==0.0
torch==2.7.0+cu128
torch==2.7.0+cu118
torchmetrics==1.7.1
torchvision==0.22.0+cu128
torchvision==0.22.0
torchvision==0.22.0+cu118
tqdm==4.67.1
tqdm==4.66.5
